m4: persist notes with room source of truth

+++ app/src/main/java/com/venkatsvision/offlinefirstsystemdesign/ui/notes/NotesUiEvent.kt
@@ -3,7 +3,7 @@ package com.venkatsvision.offlinefirstsystemdesign.ui.notes
 sealed interface NotesUiEvent {
     data class TitleChanged(val title: String) : NotesUiEvent
     data class BodyChanged(val body: String) : NotesUiEvent
-    data class EditNote(val noteId: Int) : NotesUiEvent
+    data class EditNote(val noteId: Long) : NotesUiEvent
     data object ClearEditor : NotesUiEvent
     data object SaveNote : NotesUiEvent
 }
